requirements.txt
pandas
scikit-learn
mlflow
fastapi
uvicorn
pydantic
psycopg2-binary
